+++ examples/rhine_data_bilinear.jl
@@ -23,7 +23,7 @@ if isdefined(Main, :Makie)
     y_int_pts = Vector(LinRange(spline_struct.y[1], spline_struct.y[end], n))
 
     # Get interpolated matrix
-    z_int_pts = evaluate_bicubicspline_interpolant(spline_func, x_int_pts, y_int_pts)
+    z_int_pts = evaluate_two_dimensional_interpolant(spline_func, x_int_pts, y_int_pts)
 
     plot_topography(x_int_pts, y_int_pts, z_int_pts;
                     xlabel = "E", ylabel = "N", zlabel = "H")