+++ NEWS.md
@@ -9,6 +9,7 @@ for human readability.
 
 #### Added
 
+- Added scattered radial basis function interpolation using [KernelInterpolation.jl](https://github.com/JoshuaLampert/KernelInterpolation.jl) [#73]
 - Implementation of visualization routines `plot_topography` and `plot_topography_with_interpolation_knots`
 
 #### Changed
@@ -17,5 +18,7 @@ for human readability.
 - Visualization routines rely on Makie.jl instead of PyPlot.jl [#52]
 
 #### Deprecated
+
+- Helper function `evaluate_bicubicspline_interpolant` renamed to be `evaluate_two_dimensional_interpolant` [#75]
 
 #### Removed